fix: move each hook from hooks.js into its own file within hooks folder; update imports and exports across the components
