+++ src/components/AddItems.jsx
@@ -1,7 +1,8 @@
 import { useCallback } from 'react';
-import { useStateWithStorage, normalizeItemName } from '../utils';
+import { normalizeItemName } from '../utils';
 import { addItem } from '../api';
-import TextInputElement from './TextInputElement';
+import { useStateWithStorage } from '../hooks';
+import { TextInputElement } from './TextInputElement';
 import RadioInputElement from './RadioInputElement';
 
 const daysUntilPurchaseOptions = {
