+++ src/utils/index.js
@@ -1,3 +1,2 @@
 export * from './dates';
-export * from './hooks';
 export * from './normalize';
