+++ src/views/Layout.jsx
@@ -1,6 +1,6 @@
 /* eslint-disable jsx-a11y/anchor-is-valid */
 import { Outlet, NavLink } from 'react-router-dom';
-import { useAuth, SignInButton, SignOutButton } from '../api/useAuth';
+import { useAuth, SignInButton, SignOutButton } from '../hooks/useAuth';
 // import { Home, List, ManageList } from '../views';
 
 import './Layout.css';
